add mahjong solitaire, FBI warning, compatibility
FBI Anti-Piracy Warning Seal and text
https://www.fbi.gov/investigate/white-collar-crime/piracy-ip-theft/fbi-anti-piracy-warning-seal

Several severely outdated NES emulators are still in use.  If an
NES program detects that it is incompatible with the emulator
it's running in, it could display a warning style after the UI
of the NESticle emulator.

Also a proposed set of mahjong solitaire tiles not requiring
familiarity with kanji.