blackjack: clarify hem of king's shirt
There are some jumbo index face cards in orphaned-pixels/blackjack.
A step in the hem line of each king's shirt was looking mistakenly
continuous with the inner edge of his boots.

I also put Colin's walk cycle in an animation for once and found
that a couple cels had wrong foot positioning.